myproject/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
gunicorn==20.1.0
djangorestframework==3.15.2